refactor(ts-api): move duckdb-using route tests to integration suite
Inline `#[cfg(test)] mod tests` blocks in `routes/llm_calls.rs` and
`routes/metrics.rs` pulled `ts-storage-duckdb` (and transitively the
DuckDB C++ amalgamation + 11 arrow-* crates) into the ts-api lib
unit-test binary, growing it to 53 MB / 5427 duckdb symbols. Move them
to a single new `tests/duckdb_routes.rs` integration test so the lib
unit-test binary drops to 5.8 MB / 0 duckdb symbols, while keeping the
total count of duckdb-linking binaries at 4 (consolidated into one
integration test instead of two).
