+++ server/ts-api/src/routes/llm_calls.rs
@@ -91,119 +91,3 @@ pub async fn detail(
         None => Err(ApiError::NotFound(format!("call not found: {id}"))),
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use axum::body::Body;
-    use axum::http::{Request, StatusCode};
-    use http_body_util::BodyExt;
-    use serde_json::Value;
-    use tower::ServiceExt;
-    use ts_storage_duckdb::DuckDbBackend;
-
-    use crate::{router, ApiHealthContext, ApiMetricsContext, ApiRuntimeConfigContext};
-
-    fn test_metrics_context() -> ApiMetricsContext {
-        let sys = ts_common::internal_metrics::MetricsSystem::new();
-        ApiMetricsContext {
-            pipelines: vec![],
-            global: sys.start(),
-        }
-    }
-
-    fn test_runtime_config_context() -> ApiRuntimeConfigContext {
-        ApiRuntimeConfigContext {
-            config: std::sync::Arc::new(ts_common::config::AppConfig {
-                pipelines: vec![],
-                storage: ts_common::config::StorageConfig::default(),
-                internal_metrics: ts_common::config::InternalMetricsConfig::default(),
-                api: ts_common::config::ApiConfig::default(),
-            }),
-            config_path: "test".to_string(),
-            loaded_at_ms: 0,
-            version: "test",
-        }
-    }
-
-    fn test_health_context() -> ApiHealthContext {
-        ApiHealthContext {
-            started_at_ms: 0,
-            version: "test",
-            pipelines: vec![],
-            drained: std::sync::Arc::new(std::sync::atomic::AtomicBool::new(false)),
-        }
-    }
-
-    #[tokio::test]
-    async fn invalid_status_code_returns_json_envelope() {
-        let backend = DuckDbBackend::open(":memory:").unwrap();
-        <DuckDbBackend as ts_storage::StorageBackend>::init(&backend)
-            .await
-            .unwrap();
-        let storage: std::sync::Arc<dyn ts_storage::StorageBackend> = std::sync::Arc::new(backend);
-        let app = router(
-            storage,
-            test_metrics_context(),
-            test_runtime_config_context(),
-            test_health_context(),
-            std::sync::Arc::new(vec![]),
-            ts_turn::new_active_turn_registry(),
-        );
-
-        let resp = app
-            .oneshot(
-                Request::builder()
-                    .uri("/api/llm-calls?start=0&end=1&status_code=200,abc")
-                    .body(Body::empty())
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(resp.status(), StatusCode::BAD_REQUEST);
-        assert_eq!(
-            resp.headers().get("content-type").unwrap(),
-            "application/json"
-        );
-        let bytes = resp.into_body().collect().await.unwrap().to_bytes();
-        let v: Value = serde_json::from_slice(&bytes).unwrap();
-        assert_eq!(v["code"], 1001);
-        assert!(
-            v["message"]
-                .as_str()
-                .unwrap()
-                .contains("invalid status_code: abc"),
-            "message: {}",
-            v["message"]
-        );
-    }
-
-    #[tokio::test]
-    async fn contains_params_parse() {
-        let backend = DuckDbBackend::open(":memory:").unwrap();
-        <DuckDbBackend as ts_storage::StorageBackend>::init(&backend)
-            .await
-            .unwrap();
-        let storage: std::sync::Arc<dyn ts_storage::StorageBackend> = std::sync::Arc::new(backend);
-        let app = router(
-            storage,
-            test_metrics_context(),
-            test_runtime_config_context(),
-            test_health_context(),
-            std::sync::Arc::new(vec![]),
-            ts_turn::new_active_turn_registry(),
-        );
-
-        let resp = app
-            .oneshot(
-                Request::builder()
-                    .uri("/api/llm-calls?start=0&end=1&client_ip=10.0.0.1&request_path=/v1/chat")
-                    .body(Body::empty())
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(resp.status(), StatusCode::OK);
-    }
-}
